Rename run_query, remove now unused getsecrets
